Enhancement: Use `ENTRYPOINT` instead of `CMD` for `/docker-entrypoint.sh`

+++ variants/v4.6.1-docker-alpine-3.15/docker-entrypoint.sh
@@ -1,7 +1,16 @@
 #!/bin/sh
 set -eu
-echo "Starting dockerd"
-sudo rm -fv /var/run/docker.pid
-sudo dockerd &
-echo "Starting code-server"
-exec code-server --bind-addr 0.0.0.0:8080 --disable-telemetry --disable-update-check
+
+# See: https://github.com/docker-library/official-images#consistency
+if [ "$#" -eq 0 ] || [ "${1#-}" != "$1" ]; then
+    set -- code-server "$@"
+fi
+if [ "$1" = 'code-server' ]; then
+    echo "Starting dockerd"
+    sudo rm -fv /var/run/docker.pid
+    sudo dockerd &
+
+    echo "Starting code-server"
+    exec code-server "$@"
+fi
+exec "$@"
